requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiohttp-jinja2==1.4.2
appdirs==1.4.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
bump2version==1.0.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.5
distlib==0.3.0
distro==1.4.0
filtrark==0.9.3
idna==2.8
iniconfig==1.1.1
injectark==0.7.2
ipaddr==2.2.0
jedi==0.18.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
migrark==0.6.0
modelark==0.6.4
msgpack==0.6.2
multidict==5.1.0
packaging==20.3
parso==0.8.2
pep517==0.8.2
pluggy==0.13.1
progress==1.5
psycopg2-binary==2.8.6
py==1.10.0
PyGreSQL==5.2.2
PyJWT==1.7.1
pynvim==0.4.3
pyparsing==2.4.6
pytest==6.2.3
pytest-aiohttp==0.3.0
pytest-asyncio==0.15.1
pytest-cov==2.11.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
tenark==0.8.4
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.5
uvloop==0.15.2
validark==0.2.0
webencodings==0.5.1
yarl==1.6.3
